+++ NEWS.md
@@ -1,5 +1,21 @@
 # MetaboAnnotation 0.99
 
+## Changes in 0.99.5
+
+- Add parameter `...` to `plotSpectraMirror`.
+- Definitions of "`score`", "`score_rt`" changed to be the difference
+  (with sign) between query and target m/z or retention time respectively.
+- `"ppm_error"` becomes error without sign.
+
+## Changes in 0.99.4
+
+- Add matches m/z error (variable `"ppm_error"`) to the `Matched` object
+  returned by `matchMz`.
+
+## Changes in 0.99.3
+
+- Address Herve's comments.
+
 
 # MetaboAnnotation 0.2